Bugfix multiprocessing 	modified:   data/loader/file_creator.py 	deleted:    data/loader/resources/exp_edited.tab

+++ data/loader/file_creator.py
@@ -1,5 +1,4 @@
-import argparse,subprocess, time, urllib, os
-import urllib.request
+import argparse,subprocess, time, urllib, os, warnings
 
 import multiprocessing as mp
 from multiprocessing import Queue, Process, current_process,active_children, Manager, shared_memory
@@ -8,7 +7,6 @@ from pathlib import Path
 from tqdm import tqdm
 from collections import Counter
 from multiprocessing.shared_memory import SharedMemory
-
 import pandas as pd
 import numpy as np
 
@@ -48,20 +46,32 @@ cmd_line.add_argument(
 cmd_line.add_argument(
     '--nworkers',
     '-n',
-    default=6,
+    default=8,
     type= int
 )
 cmd_line.add_argument(
-    "--chunk_size",
+    "--chunksize",
     '-s',
     type = int,
-    default=100000
+    default=500000
 )
 cmd_line.add_argument(
     "--assembly_threshold",
     '-t',
     type=str,
     default='05'
+)
+cmd_line.add_argument(
+    "--reload",
+    '-r',
+    type=int,
+    default=0
+)
+cmd_line.add_argument(
+    "--download",
+    "-d",
+    type=int,
+    default=0
 )
 """End of block for cmd line arguments"""  
 
@@ -70,32 +80,31 @@ def writer(que: Queue, done:Queue, file_names: dict):
     pause = 0
     is_working = True
     while is_working:
-        if not que.empty(): 
+        if not que.empty():# working cycle 
             pause = 0   
             SM_meta, df, lines = que.get()
-            try:
+            try: # this check for Done
                 ShMem = shared_memory.SharedMemory(name=df)
                 list_copy = np.ndarray(SM_meta["shape"], dtype=SM_meta["dtype"], buffer=ShMem.buf)
-                try:
-                    [open(file_names[line[3]], "+a").write('\t'.join(line)) for line in lines]
-                except:
-                         pass
-            except:
+                for id,values in lines.items():
+                    try: # writing only experiment list tab experiments
+                        open(file_names[id],'+a').writelines(['\t'.join(line)+'\n' for line in list_copy[[values]][0]])
+                    except KeyError:
+                        pass 
+                ShMem.close()
+                done.put(df)   
+            except Exception as e:
                 if SM_meta == "Done":
                     is_working = False
                     done.put("Done")
-                pass
-                       #print(f"excepted {to_write[3]} ")
-                
-            ShMem.close()
-            done.put(df)            
+                else:
+                    raise e
         else:
             pause +=1
             time.sleep(0.1)
             if pause > 100:
                 print(f"{current_process().name} paused to {pause * 0.1}s")
-
-    print(f"{current_process().name} stopped")
+    #print(f"{current_process().name} stopped")
     
 
 def worker_file_creator(df,tasks, path, check:bool, file_dict: dict):
@@ -109,17 +118,18 @@ def worker_file_creator(df,tasks, path, check:bool, file_dict: dict):
                 if not Path.is_file(actual_path):
                     file_dict["No_file"] += 1
             else:
-                #with open(actual_path, 'w') as f:
-                #    f.close()
-                    file_dict[file[0]] = actual_path
+                open(actual_path, 'w+').close()
+                file_dict[file[0]] = actual_path
 
 
-def ExpListProcessing(data: pd.DataFrame, path: str, n_workers: int, check: bool, file_dict: dict):
+def ExpListProcessing(data: pd.DataFrame, path: Path, n_workers: int, check: bool, file_dict: dict):
     """ Ortem`s functions"""
     tasks = Queue()
     for batch_start in range(0, len(data) - 1000, 1000):
         tasks.put((batch_start, batch_start + 1000))
     tasks.put((len(data) - 1000, len(data)))
+    if check:
+        file_dict["No_file"] = 0
     
     procs = [Process(target=worker_file_creator, args=(data, tasks,path, check, file_dict )) for _ in range(n_workers) ]
 
@@ -127,22 +137,20 @@ def ExpListProcessing(data: pd.DataFrame, path: str, n_workers: int, check: bool
     [p.join()   for p in procs]
     return 0
     
-def Put2QandSM_Process(df:pd.DataFrame, lines_bufer: dict, writers_que: list):
+def Put2QandSM_Process(df:pd.DataFrame, lines_bufer: list, writers_que: list):
     """Function for put and free buffer from __main__ process to child`s Q"""
-    Df2List = np.array(df.values.tolist())
+    Df2List = df.to_numpy(dtype=str)
     ShMem = shared_memory.SharedMemory(create=True, size=Df2List.nbytes)
     SM_list = {"shape":Df2List.shape, "dtype": Df2List.dtype}
     l_copy = np.ndarray(Df2List.shape, dtype=Df2List.dtype, buffer=ShMem.buf)
     l_copy[:] = Df2List[:]
     pause = 0
-    #print("w8")
     while (not all(q.empty() for q in writers_que)):
         pause += 1
         time.sleep(1)
         pass
-    for buf in lines_bufer:
-        writers_que[buf].put((SM_list,ShMem.name,lines_bufer[buf]))
-        #print((SM_list,ShMem.name,lines_bufer[buf]))
+    for buf,lines in enumerate(lines_bufer):
+        writers_que[buf].put((SM_list,ShMem.name,lines))
     ShMem.close()
     #print(f"this iter was waiting to {pause*0.01}s")
 
@@ -154,33 +162,44 @@ def FreeTheMemmory_Process(DoneQ : Queue, n_writers: int):
         while not DoneQ.empty():
             done_counter[done_que.get()] +=1
         for mem_buf in done_counter:
-                if done_counter[mem_buf] == n_writers:
-                    ShMem = shared_memory.SharedMemory(name=mem_buf)
-                    ShMem.unlink()
-                    done_counter[mem_buf] = 0
-                    #print(f"free {mem_buf}")
-        if  done_counter["Done"] != 0:
-            for mem_buf in done_counter:
-                 if done_counter[mem_buf] != 0:
-                    ShMem = shared_memory.SharedMemory(name=mem_buf)
-                    ShMem.unlink()
-            is_working = False
+            if done_counter[mem_buf] == n_writers and not ("Done" in mem_buf):
+                ShMem = shared_memory.SharedMemory(name=mem_buf)
+                ShMem.unlink()
+                done_counter[mem_buf] = 0
+                #print(f"free {mem_buf}")
+            elif done_counter[mem_buf] == n_writers:
+                is_working = False
         time.sleep(1)
     print(f"{current_process().name}(GC) stopped")
+
+def DownloadChipAtlasFile(Dir: Path, assembly: str, threshold: str) -> subprocess.Popen:
+    os.mkdir(Dir) 
+    import urllib.request,urllib
+    urllib = getattr(urllib, 'request', urllib)
+    Path2File = Dir / f"allPeaks_light.{args.assembly}.{args.assembly_threshold}.bed.gz"
+    eg_link = f"https://chip-atlas.dbcls.jp/data/{args.assembly}/allPeaks_light/allPeaks_light.{args.assembly}.{args.assembly_threshold}.bed.gz"
+    with TqdmUpTo(unit='B', unit_scale=True, unit_divisor=1024, miniters=1,
+                  desc=eg_link.split('/')[-1]) as t:  # all optional kwargs
+        urllib.urlretrieve(eg_link, filename= Path2File,
+                           reporthook=t.update_to, data=None)
+        t.total = t.n 
+    return subprocess.Popen(["gunzip",Path2File])    
+
         
 
 if __name__ == '__main__':
     args = cmd_line.parse_args()
+    SubporocessHub = None
     NWORKERS = args.nworkers
-    NWRITERS = 20
-    file_dick = Manager().dict()
+    Files = Manager().dict()
     WORKING_DIR = Path(f"./resources/{args.assembly}")
-    chunk_size = 500000
+    chunk_size = args.chunksize
     
 
     if not Path('./resources/experimentList.tab').exists(): # if user has no ungz list tab - ungz it
-        proc = subprocess.Popen(["gunzip",Path('./resources/experimentList.tab.gz')])
-        proc.wait()
+        SubporocessHub = subprocess.Popen(["gunzip",Path('./resources/experimentList.tab.gz')])
+        SubporocessHub.wait()
+        SubporocessHub = None
 
     # read ExpList df
     ExpList = pd.read_csv(Path("./resources/experimentList.tab"),
@@ -193,30 +212,25 @@ if __name__ == '__main__':
         inplace=False, regex=True
         )
     
+    if args.reload and args.download:
+        SubporocessHub = subprocess.Popen(["rm","-rf", WORKING_DIR])
+        SubporocessHub.wait()
+        SubporocessHub = None
 
-    if WORKING_DIR.exists():
-        #os.mkdir(WORKING_DIR) 
-        """This condition for first setup with download assembly file from CA DB"""
-        #urllib = getattr(urllib, 'request', urllib)
-        Path2File = WORKING_DIR / f"allPeaks_light.{args.assembly}.{args.assembly_threshold}.bed.gz"
-        # eg_link = f"https://chip-atlas.dbcls.jp/data/{args.assembly}/allPeaks_light/allPeaks_light.{args.assembly}.{args.assembly_threshold}.bed.gz"
-        # with TqdmUpTo(unit='B', unit_scale=True, unit_divisor=1024, miniters=1,
-        #             desc=eg_link.split('/')[-1]) as t:  # all optional kwargs
-        #     urllib.urlretrieve(eg_link, filename= Path2File,
-        #                     reporthook=t.update_to, data=None)
-        #     t.total = t.n
-
-        #proc = subprocess.Popen(["gunzip",Path2File])     
-
+    if (not WORKING_DIR.exists()) or args.download: SubporocessHub = DownloadChipAtlasFile(WORKING_DIR,args.assembly, args.assembly_threshold)
+   
+    if args.reload or args.check:
         """Create files for each id experiment"""
-        ExpListProcessing(ExpList, n_workers=12,path = WORKING_DIR, check = False, file_dict = file_dick)
+        ExpListProcessing(ExpList, n_workers=NWORKERS,path = WORKING_DIR, check = args.check, file_dict = Files)
+    
+    if args.check and Files["No_file"]:
+        exit("Need reload")
 
-        #proc.wait()
+    if SubporocessHub: SubporocessHub.wait()
+
+    if args.reload:
         Path2File = WORKING_DIR / f"allPeaks_light.{args.assembly}.{args.assembly_threshold}.bed"
-        LinesInDock = str(subprocess.check_output(['wc','-l',Path2File]))
-        LinesInDock = int(LinesInDock[2:LinesInDock.find(' ')])
-
-        print(LinesInDock)
+        SubporocessHub = subprocess.Popen(['wc','-l',Path2File], stdout= subprocess.PIPE)
 
         iter_df = pd.read_csv(
                 Path2File,
@@ -225,44 +239,46 @@ if __name__ == '__main__':
                 header= None
             )
 
-        lines_buf = {key:[] for key in range(NWRITERS)}# buffer for each chunk
+        lines_buf = [{} for key in range(NWORKERS)]# buffer for each chunk
         writer_list = []# buffer for writers
 
         """Creating queues"""
-        workers_ques = [ Queue() for i in range(NWRITERS) ]
+        workers_ques = [ Queue() for i in range(NWORKERS) ]
         done_que = Queue()
         
         # setup and run checker to free memory
-        Checker = Process(target = FreeTheMemmory_Process, args = (done_que,NWRITERS,))
+        Checker = Process(target = FreeTheMemmory_Process, args = (done_que,NWORKERS,))
         Checker.start()
 
-        #print("exala")
+        #cycle =0
+        SubporocessHub.wait()
+        LinesInDock = str(SubporocessHub.communicate())
+        LinesInDock = int(LinesInDock[3:LinesInDock.find(' ')])
+
         for chunk in tqdm(iter_df, total = int(LinesInDock/chunk_size)):
             """W8 till each process take his part from writers_Q"""
             for index,id_exp in enumerate(chunk[3]):
                 """ Putting lines for each writer to work in buffer for  writersQ"""
-                q_num = int(id_exp[3:])%NWRITERS
-                lines_buf[q_num].append(index)
+                q_num = int(id_exp[3:])%NWORKERS
+                try:
+                    lines_buf[q_num][id_exp] += index,
+                except KeyError:
+                    lines_buf[q_num][id_exp] = index,
             if not writer_list:
                  # Setup and run writers        
-                for i in range(NWRITERS):
-                    p = Process(target=writer, args=(workers_ques[i], done_que, file_dick,))
+                for i in range(NWORKERS):
+                    p = Process(target=writer, args=(workers_ques[i], done_que, Files,))
                     writer_list.append(p)
                     p.start()
             """without sending package to writerQ, writer cannot take smth from chunkQ"""
             Put2QandSM_Process(chunk,lines_buf,workers_ques)
-            lines_buf = {key:[] for key in range(NWRITERS)}# refresh buffer
+            lines_buf = [{} for key in range(NWORKERS)]# refresh buffer
+
            
-        
+        time.sleep(2)
         [q.put(("Done",0,0)) for q in workers_ques]
         [wr.join()  for wr  in writer_list]
         Checker.join()
-        #subprocess.run(f"rm ./resources/allPeaks_light.{args.assembly}.{args.assembly_threshold}.bed.gz")
+        subprocess.Popen(["rm", Path2File])
     
-    elif args.check:
-        file_dick["No_file"] = 0 
-        result = ExpListProcessing(df, n_workers=8, check = False, file_dict = file_dick)
-        print(file_dick["No_file"])
-    
-    else:
-        print("Already satisfaied")    
+    exit("Done")
